% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_themes.R
\name{add_theme}
\alias{add_theme}
\title{Add a theme to the tableHTML}
\usage{
add_theme(tableHTML, theme = c("scientific", "rshiny-blue"))
}
\arguments{
\item{theme}{Pick one of the provided themes. These can still be modified by extra css. Choices
are: scientific, rstudio-blue. Column widths are not provided when you select a theme.
Please use the width argument for column widths.}
}
\value{
A tableHTML object.
}
\description{
\code{add_theme} will add a theme to tableHTML
}
\details{
\code{add_theme} will add a theme to tableHTML.
}
\examples{
tableHTML(mtcars,
          rownames = FALSE,
          widths = c(140, rep(50, 11)),
          row_groups = list(c(10, 10, 12), c('Group 1', 'Group 2', 'Group 3')),
          second_headers = list(c(3, 4), c('col1', 'col2'))) \%>\%
  add_theme('scientific')

tableHTML(mtcars, widths = c(140, rep(50, 11))) \%>\%
  add_theme ('rshiny-blue')

}
