% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logo.R
\name{create_logo}
\alias{create_logo}
\title{Generate package's logo}
\usage{
create_logo(save = TRUE, format = "html",
  file = "tableHTML_logo.html", complete_html = TRUE, ...)
}
\arguments{
\item{save}{A boolean when set to TRUE the logo will be saved in
the specified format.}

\item{format}{A character string to specify the format of the output,
it accepts \code{'html'}, \code{'png'}, or \code{'jpeg'}.
Default is \code{'html'}.}

\item{file}{A character string to specify the name and path to the new file.
Should end with \code{'.html'}, \code{'.png'}, or \code{'.jpeg'},
depending on the selected format.}

\item{complete_html}{Either TRUE or FALSE. Defaults to FALSE. If TRUE then the <html> and <body>
tags are also added in the file.}

\item{...}{Further parameters to pass to webshot.}
}
\value{
The logo of the tableHTML package as a tableHTML object.
The output will be shown in the Viewer pane, and can be saved
either as an image or as an HTML.
}
\description{
\code{create_logo} will create and generate the package's logo.
}
\details{
\code{create_logo} will create and generate the package's logo.
}
\examples{
create_logo(save = FALSE)

create_logo(format = 'png', file = '~/tableHTML_export.png')

}
