% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_css.R
\name{make_css}
\alias{make_css}
\title{Create a css file}
\usage{
make_css(...)
}
\arguments{
\item{...}{css style definitions}
}
\value{
A css file.
}
\description{
\code{make_css} will create a css file which can also be used in shiny
}
\details{
\code{make_css} will create a css file which can also be used in shiny
}
\examples{
make_css(list('table', c('text-align', 'font-size'), c('center', '20px')),
         list('th', c('text-align', 'font-size'), c('center', '20px')))

}

