% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_css_conditional_format.R
\name{make_css_colour_rank_theme}
\alias{make_css_colour_rank_theme}
\title{Get css properties for custom colour rank theme}
\usage{
make_css_colour_rank_theme(column_data, colors, decreasing = FALSE,
  same_scale = TRUE)
}
\arguments{
\item{column_data}{A named list of vectors of values that are in a tableHTML column which
should be mapped to a colour palette.}

\item{colors}{colors to interpolate; must be a valid argument to
    \code{\link{col2rgb}()}.}

\item{decreasing}{logical.  Should the sort order be increasing or
    decreasing? For the \code{"radix"} method, this can be a vector of
    length equal to the number of arguments in \code{\dots}. For the
    other methods, it must be length one.}

\item{same_scale}{This logical flag indicates whether all of the columns the css should be applied to 
are on the same scale. If TRUE, the same value in all columns will have the same colour applied to it.
If FALSE, the colours will be assigned between the columns min and max.}
}
\value{
A list of css properties
}
\description{
\code{make_css_colour_rank_theme} will create a list of css properties needed for custom conditional formatting.
}
\details{
\code{make_css_colour_rank_theme} will add conditional css to a tableHTML's columns. \code{add_conditional_css_column} will only
  add css to the columns without the headers or second headers (i.e. it only affects the td tag
  internally and not the th tag). If you want to add css to the headers or second headers please
  use \code{add_css_header} or \code{add_css_second_header}.
}
\examples{

tableHTML <- tableHTML(mtcars)

css <- make_css_colour_rank_theme(list(mpg = mtcars$mpg), c("orange", "yellow","springgreen","royalblue"))

tableHTML \%>\% add_css_conditional_column(conditional = "colour_rank",
                                        colour_rank_theme = "Custom", colour_rank_css = css, column = 1)

}
