% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_even.R
\name{odd}
\alias{even}
\alias{odd}
\title{Get the odd or even numbers from a numeric vector}
\usage{
odd(vec)

even(vec)
}
\arguments{
\item{vec}{A numeric atomic vector.}
}
\value{
A numeric atomic vector with the odd / even numbers
}
\description{
Get the odd or even numbers from a numeric vector
}
\details{
\code{odd} will extract the odd numbers from a vector.\cr \code{even} will extract the 
  even numbers from a vector
}
\examples{
odd(1:10)
even(1:10)

}

