% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableHTML2raster.R
\name{tableHTML_to_raster}
\alias{tableHTML_to_raster}
\title{Convert a tableHTML image into a raster object}
\usage{
tableHTML_to_raster(tableHTML_image, ...)
}
\arguments{
\item{tableHTML_image}{A tableHTML_image object created from the \code{tableHTML_to_image}
function or a 3D numeric array representing an image.}

\item{...}{Arguments to pass to \link[grid]{rasterGrob}.}
}
\value{
A raster object
}
\description{
\code{tableHTML_to_raster} converts the tableHTML image into a raster object.
}
\details{
The main rational behind this function is to convert the image into a raster object in
order to use alongside ggplot. Check the examples.
}
\examples{
\dontrun{
#create the raster object
raster_object <- iris \%>\%
 head() \%>\%
 tableHTML(widths = c(rep(120, 5)), rownames = FALSE) \%>\%
 add_theme('scientific') \%>\%
 tableHTML_to_image() \%>\%
 tableHTML_to_raster()

#plot as part of ggplot
library(ggplot2)
ggplot() +
 theme_void() +
 annotation_custom(raster_object, xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf)

#combine in a grid
library(grid)
library(cowplot)
library(dplyr)
gg1 <- ggplot(iris, aes(Sepal.Length, Petal.Length, color = Species)) + geom_point()
gg2 <- ggplot() +
 theme_void() +
 annotation_custom(raster_object, xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf)

cowplot::plot_grid(gg1, gg2, nrow = 2, ncol = 1, rel_heights = c(3, 1))
}

}
