% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableHTML.R
\name{tableHTML}
\alias{print.tableHTML}
\alias{tableHTML}
\title{Create an easily css-ible HTML table}
\usage{
tableHTML(obj, rownames = TRUE, class = paste0("table_",
  deparse(substitute(obj))), widths = NULL, second_header = NULL,
  row_groups = NULL, caption = NULL, footer = NULL, border = 1,
  theme = c("default", "scientific", "rshiny-blue"))

\method{print}{tableHTML}(x, viewer = TRUE, ...)
}
\arguments{
\item{obj}{Needs to be a data.frame or a matrix or an arbitrary object that has the 
data.frame class and can be coersible to a data.frame (e.g data.table, tbl, etc.)}

\item{rownames}{Can be TRUE or FALSE. Defaults to TRUE. Whether the obj's rownames
will be inlcuded.}

\item{class}{Character string. Specifies the table's class. Convinient if you have multiple
tables. Default is table_<data_frame_name>.}

\item{widths}{Needs to be a numeric atomic vector with the column widths. Widths are in pixels.}

\item{second_header}{A list of two elements of the same length. The first element will contain
the column spans (i.e. a numeric atomic vector) whereas the second element will contain the
names (i.e. a character atomic vector). See the examples for more info. Defauls to NULL.}

\item{row_groups}{A list of two elements of the same length. The first element will contain
the row spans (i.e. a numeric atomic vector) whereas the second element will contain the
names (i.e. a character atomic vector). See the examples for more info. Defauls to NULL.}

\item{caption}{Character string. The table's caption.}

\item{footer}{Character string. The table's footer. This gets added below the table and it
should not be confused with tfooter.}

\item{border}{An integer. Specifies the border of the table. Defaults to 1. 0 removes borders 
from the table. The higher the number the thicker the table's outside border.}

\item{theme}{Pick one of the provided themes. These can still be modified by extra css. Choices 
are: default, scientific. column widths are not provided when you select a theme. Please use
the width argument for column widths. Defaults to 'default' i.e. no css included.}

\item{x}{A tableHTML object created from the \code{tableHTML} function.}

\item{viewer}{TRUE or FALSE. Defaults to TRUE. Whether or not to render the HTML table. If 
you are working on Rstudio (interactively) the table will be printed or Rstudio's viewer. 
If you are working on Rgui (interactively) the table will be printed on your default browser.
If you set this to FALSE the HTML code will be printed on screen.}

\item{...}{Optional arguments to print.}
}
\value{
A tableHTML object. Printing the table will result in rendering it in R studio's viewer
        with the print.tableHTML method if using Rstudio otherwise it will use the default 
        browser. Use \code{str(tableHTML)} to view the actual html code.
}
\description{
The purpose of \code{tableHTML} is to create easily css-ible HTML tables
that are compatible with R shiny. The exported HTML table will contain separate ids
or classes for headers, columns, second headers (if any) and the table itself 
(in case you have multiple tables) in order to create a 
complex css file very easily. ids and classes are explained in detail in 
the details section.
}
\details{
\code{tableHTML} will create an HTML table with defined ids and classes for rows and columns.
In particular:
\itemize{
  \item \strong{Table:} Will get the class from the class argument in the function. 
      The final class will be of the form table_<class_name>.
  \item \strong{Columns:} Will get an id which will be the same as the column name
  \item \strong{Headers:} Will get an id of the form header_<header index>. 
      For example the first header will have the id header_1, the second header 
      will have header_2 and so on.
  \item \strong{Second_Header:} Will get an id of the form overheader_<second header index>.  
      For example the first second_header will have the id overheader_1, the second header  
      will have overheader_2 and so on.
}

Notice that rows do not get a specific id or class as they are easily accessible from css 
anyway.
}
\examples{
tableHTML(mtcars)
tableHTML(mtcars, rownames = FALSE)
tableHTML(mtcars, class = 'table1')
tableHTML(mtcars, second_header = list(c(3, 4, 5), c('col1', 'col2', 'col3')))
tableHTML(mtcars, 
          widths = c(rep(50, 6), rep(100, 6)), 
          second_header = list(c(3, 4, 5), c('col1', 'col2', 'col3')))
tableHTML(mtcars, caption = 'This is a caption', footer = 'This is a footer')
tableHTML(mtcars, 
          row_groups = list(c(10, 10, 12), c('Group 1', 'Group 2', 'Group 3')), 
          widths = c(200, rep(50, 5), rep(100, 6)), 
          rownames = FALSE)
tableHTML(mtcars, 
          rownames = FALSE, 
          widths = c(120, rep(50, 11)),
          row_groups = list(c(10, 10, 12), c('Group 1', 'Group 2', 'Group 3')),
          second_header = list(c(3, 4, 5), c('col1', 'col2', 'col3')))
tableHTML(mtcars, 
          rownames = FALSE, 
          widths = c(120, rep(50, 11)),
          row_groups = list(c(10, 10, 12), c('Group 1', 'Group 2', 'Group 3')),
          second_header = list(c(3, 4), c('col1', 'col2')), 
          theme = 'scientific')
tableHTML(mtcars, theme = 'rshiny-blue', widths = c(140, rep(50, 11)))

}

